refactor(date): parsetimestring を lib/date に移動
calendar/interaction/time-math.ts に閉じていた "HH:mm" parser を
lib/date/timeString.ts に切り出し、@/lib/date barrel から再 export。

理由: project a-4 で entry feature の inspector 入力 component が同 parser を
再利用するため。entry → calendar の deep import は feature boundary 違反
（entry layer 1 / calendar layer 2）。pure parser を lib に上げて両 feature
から参照可能にする。

formathhmm も併設し、{hour, minute} → "HH:mm" の 0 padding 整形を共通化。
