+++ src/features/calendar/interaction/time-math.test.ts
@@ -1,10 +1,11 @@
 import { describe, expect, it } from 'vitest';
 
+import { parseTimeString } from '@/lib/date';
+
 import {
   addMinutesToTime,
   DEFAULT_SNAP_INTERVAL,
   formatTimeString,
-  parseTimeString,
   pixelsToTime,
   pixelsToTimeUnsnapped,
   snapDeltaToGrid,
